refactor!: move redb to module on safebrowsing-db crate
