--- safebrowsing-db-redb/CHANGELOG.md
@@ -1,14 +1,0 @@
-# Changelog
-
-All notable changes to this project will be documented in this file.
-
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
-## [Unreleased]
-
-## [0.1.0](https://github.com/JadedBlueEyes/safebrowsing-rs/releases/tag/safebrowsing-db-redb-v0.1.0) - 2025-08-07
-
-### Other
-
-- Initial commit
