--- safebrowsing-db-redb/Cargo.toml
@@ -1,36 +1,0 @@
-[package]
-name = "safebrowsing-db-redb"
-version.workspace = true
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-description = "Redb-based persistent database backend for Safe Browsing API"
-repository.workspace = true
-keywords.workspace = true
-categories.workspace = true
-
-[dependencies]
-tokio = { workspace = true, features = ["sync", "time"] }
-thiserror.workspace = true
-tracing.workspace = true
-serde = { workspace = true, features = ["derive"] }
-serde_json.workspace = true
-async-trait.workspace = true
-bytes.workspace = true
-redb = "2.1"
-dirs = "5.0"
-
-# Safe browsing crates
-safebrowsing-api.workspace = true
-safebrowsing-db.workspace = true
-safebrowsing-hash.workspace = true
-safebrowsing-proto.workspace = true
-
-[dev-dependencies]
-tokio-test.workspace = true
-tempfile.workspace = true
-tracing-subscriber.workspace = true
-
-[[example]]
-name = "basic_usage"
-required-features = []
